stats: add kotlin Counter interface
